add most of the basic infrastructure
