+++ .flake8
@@ -1,0 +1,31 @@
+[flake8]
+ignore =
+    E203
+    E266
+    E501
+    W503
+    F403
+    F401
+    E231
+max-line-length = 88
+max-complexity = 18
+select = B,C,E,F,W,T4,B9,D
+exclude =
+    setup.py
+rst-roles =
+    class
+    func
+    mod
+    data
+    const
+    meth
+    attr
+    exc
+    obj
+rst-directives =
+    note
+    warning
+    versionadded
+    versionchanged
+    deprecated
+    seealso
