+++ .pre-commit-config.yaml
@@ -1,0 +1,46 @@
+exclude: '^docs/conf.py'
+
+repos:
+- repo: git://github.com/pre-commit/pre-commit-hooks
+  rev: v2.5.0
+  hooks:
+  - id: trailing-whitespace
+  - id: check-added-large-files
+  - id: check-ast
+  - id: check-json
+  - id: check-merge-conflict
+  - id: check-xml
+  - id: check-yaml
+  - id: debug-statements
+  - id: end-of-file-fixer
+  - id: requirements-txt-fixer
+  - id: check-merge-conflict
+  - id: mixed-line-ending
+    args: ['--fix=no']
+-   repo: https://gitlab.com/pycqa/flake8
+    rev: '3.8.1'  # pick a git hash / tag to point to
+    hooks:
+    -   id: flake8
+        additional_dependencies:
+          - flake8-quotes
+          - flake8-comprehensions
+          - flake8-builtins
+          - flake8-eradicate
+          - pep8-naming
+          - flake8-docstrings
+          - flake8-rst-docstrings
+          - flake8-rst
+          - flake8-copyright
+          - flake8-ownership
+-   repo: https://github.com/psf/black
+    rev: 19.10b0
+    hooks:
+    - id: black
+-   repo: https://github.com/pre-commit/pygrep-hooks
+    rev: v1.5.1
+    hooks:
+      - id: rst-backticks
+- repo: https://github.com/timothycrosley/isort
+  rev: 5.0.5
+  hooks:
+  - id: isort
