+++ CHANGELOG.rst
@@ -1,0 +1,14 @@
+Releases
+========
+
+dev
+---
+
+Fixed
+~~~~~
+
+Changed
+~~~~~~~
+
+Added
+~~~~~
