--- README.md
@@ -1,3 +1,0 @@
-# pspec_likelihood
-
-**A small but powerful interface to generate theoretical likelihoods from `UVPSpec` objects.**
