+++ README.rst
@@ -1,0 +1,38 @@
+pspec_likelihood
+================
+
+**A small but powerful interface to generate theoretical likelihoods from ``UVPSpec`` objects.**
+
+
+.. image:: https://github.com/hera-team/pspec_likelihood/workflows/Tests/badge.svg
+    :target: https://github.com/steven-murray/hmf
+.. image:: https://codecov.io/gh/hera-team/pspec_likelihood/branch/master/graph/badge.svg
+    :target: https://codecov.io/gh/steven-murray/hmf
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/psf/black
+
+
+Full Documentation
+------------------
+`Read the docs. <http://pspec_likelihood.readthedocs.org>`_
+
+Features
+--------
+* Feature 1
+* Feature 2
+
+
+Quickstart
+----------
+Import like this...
+
+
+Versioning
+----------
+From v0.1.0, ``pspec_likelihood`` will be using strict semantic versioning, such that increases in
+the **major** version have potential API breaking changes, **minor** versions introduce
+new features, and **patch** versions fix bugs and other non-breaking internal changes.
+
+If your package depends on ``pspec_likelihood``, set the dependent version like this::
+
+    pspec_likelihood>=0.1<1.0
