+++ setup.cfg
@@ -1,0 +1,91 @@
+# This file is used to configure your project.
+# Read more about the various options under:
+# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+
+[metadata]
+name = pspec_likelihood
+description = A likelihood based on UVPSpec objects
+author = HERA-Team
+license = mit
+long_description = file: README.rst
+long_description_content_type = text/x-rst; charset=UTF-8
+url = https://github.com/hera-team/pspec_likelihood
+project_urls =
+    Documentation = https://pspec_likelihood.readthedocs.org
+# Change if running only on Windows, Mac or Linux (comma-separated)
+platforms = any
+# Add here all kinds of additional classifiers as defined under
+# https://pypi.python.org/pypi?%3Aaction=list_classifiers
+classifiers =
+    Development Status :: 1 - Planning
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Intended Audience :: Science/Research
+    License :: OSI Approved
+    License :: OSI Approved :: MIT License
+    Natural Language :: English
+    Topic :: Scientific/Engineering :: Physics
+    Topic :: Scientific/Engineering :: Astronomy
+
+[options]
+zip_safe = False
+packages = find:
+include_package_data = True
+package_dir =
+    =src
+install_requires =
+    hera_pspec @ git+git://github.com/hera-team/hera_pspec
+    attrs
+
+[options.packages.find]
+where = src
+exclude =
+    tests
+
+[options.extras_require]
+docs =
+    Sphinx>=1.7.5
+    numpydoc>=0.8.0
+    nbsphinx
+tests =
+    coverage>=4.5.1
+    pytest>=3.5.1
+    pytest-cov>=2.5.1
+    pre-commit
+dev =
+    Sphinx>=1.7.5
+    numpydoc>=0.8.0
+    nbsphinx
+    coverage>=4.5.1
+    pytest>=3.5.1
+    pytest-cov>=2.5.1
+    pre-commit
+    setuptools_scm
+
+[tool:pytest]
+addopts =
+    --cov pspec_likelihood --cov-report term-missing
+    --verbose
+norecursedirs =
+    dist
+    build
+    .tox
+testpaths = tests
+
+[aliases]
+dists = bdist_wheel
+
+[bdist_wheel]
+# Use this option if your package is pure-python
+universal = 1
+
+[build_sphinx]
+source_dir = docs
+build_dir = build/sphinx
+
+[devpi:upload]
+# Options for the devpi: PyPI server and packaging tool
+# VCS export must be deactivated since we are using setuptools-scm
+no-vcs = 1
+formats = bdist_wheel
