+++ src/pspeclikelihood/__init__.py
@@ -1,0 +1,3 @@
+"""The PSpec Likelihood Package."""
+from . import likelihood
+from .likelihood import PSpecLikelihood
